# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anglebert Team Employee Manager API",
        "description": "Advanced employee management system (Anglebert Team) featuring Sanctum authentication, attendance tracking, and reporting.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/attendance": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "List attendance records",
                "operationId": "34b5a0f9b766a9e4cb3440357d371351",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of attendance records",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Attendance"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/attendance/check-in": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Record employee check-in",
                "operationId": "9379095aac0cd891a8397007afa6d03a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Check-in recorded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attendance"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/attendance/check-out": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Record employee check-out",
                "operationId": "3f3748aaa8b4a063a7c01f8167f71620",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check-out recorded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Attendance"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Open attendance record not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/reports/attendance/daily/pdf": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Download daily attendance report as PDF",
                "operationId": "e1f18a2e0ad47d3787df6acc7fc35d4f",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/reports/attendance/daily/excel": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Download daily attendance report as Excel",
                "operationId": "88c293640fae6980a49a7cd4488bd8ca",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/reports/attendance/daily/email": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Send daily attendance report (PDF & Excel) to your email",
                "operationId": "4da0af1508b1abbc0acd5058e0f4c02a",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report queued for delivery",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The report has been queued for delivery to your email."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Attendance": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "employee_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "check_in_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "check_out_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}